ppp2/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
psycopg[binary]==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
